Refactor _actions/_top.py into two files
One file will now contain pool specific actions,
while the other holds top-level actions.
